+++ ZoomIn/ZoomIn/StockControl/Curve.cs
@@ -6,7 +6,7 @@ namespace ZoomIn.StockControl
 {
     public class Curve : StockShapeBase
     {
-        public void CreateGeometry(double[] values, int gap, int width)
+        public void CreateGeometry(double[] values, double gap, double width)
         {
             var geometryGroup = new GeometryGroup();
             var streamGeometry = new StreamGeometry();
